re-use the platform authorization entity, which meant its retrieval via the standalone service had to be async
