+++ src/platform/platfrom/platform.service.authorization.ts
@@ -20,11 +20,18 @@ import { InnovationHubService } from '@domain/innovation-hub';
 import { InnovationHubAuthorizationService } from '@domain/innovation-hub/innovation.hub.service.authorization';
 import {
   CREDENTIAL_RULE_TYPES_PLATFORM_ACCESS_GUIDANCE,
+  CREDENTIAL_RULE_TYPES_PLATFORM_ADMINS,
+  CREDENTIAL_RULE_TYPES_PLATFORM_ANY_ADMIN,
   CREDENTIAL_RULE_TYPES_PLATFORM_FILE_UPLOAD_ANY_USER,
+  CREDENTIAL_RULE_TYPES_PLATFORM_GRANT_GLOBAL_ADMINS,
+  CREDENTIAL_RULE_TYPES_PLATFORM_READ_REGISTERED,
+  POLICY_RULE_PLATFORM_CREATE,
 } from '@common/constants';
 import { StorageAggregatorAuthorizationService } from '@domain/storage/storage-aggregator/storage.aggregator.service.authorization';
 import { ICredentialDefinition } from '@domain/agent/credential/credential.definition.interface';
 import { OrganizationService } from '@domain/community/organization/organization.service';
+import { AuthorizationPolicyRulePrivilege } from '@core/authorization/authorization.policy.rule.privilege';
+import { RelationshipNotFoundException } from '@common/exceptions/relationship.not.found.exception';
 
 @Injectable()
 export class PlatformAuthorizationService {
@@ -45,6 +52,7 @@ export class PlatformAuthorizationService {
   async applyAuthorizationPolicy(): Promise<IPlatform> {
     const platform = await this.platformService.getPlatformOrFail({
       relations: {
+        authorization: true,
         library: {
           innovationPacks: true,
         },
@@ -52,17 +60,29 @@ export class PlatformAuthorizationService {
       },
     });
 
-    const platformAuthorizationPolicy =
-      this.platformAuthorizationPolicyService.getPlatformAuthorizationPolicy();
+    if (!platform.authorization)
+      throw new RelationshipNotFoundException(
+        `Unable to load entities for platform: ${platform.id} `,
+        LogContext.PLATFORM
+      );
 
+    platform.authorization = await this.authorizationPolicyService.reset(
+      platform.authorization
+    );
     platform.authorization =
-      this.authorizationPolicyService.cloneAuthorizationPolicy(
-        platformAuthorizationPolicy
+      this.platformAuthorizationPolicyService.inheritRootAuthorizationPolicy(
+        platform.authorization
       );
     platform.authorization.anonymousReadAccess = true;
+    platform.authorization = await this.appendCredentialRules(
+      platform.authorization
+    );
+
+    const privilegeRules = this.createPlatformPrivilegeRules();
     platform.authorization =
-      await this.appendCredentialRulesInteractiveGuidance(
-        platform.authorization
+      this.authorizationPolicyService.appendPrivilegeAuthorizationRules(
+        platform.authorization,
+        privilegeRules
       );
 
     // Cascade down
@@ -71,6 +91,20 @@ export class PlatformAuthorizationService {
     );
 
     return await this.platformRepository.save(platformPropagated);
+  }
+
+  private async appendCredentialRules(
+    authorization: IAuthorizationPolicy
+  ): Promise<IAuthorizationPolicy> {
+    const credentialRules = this.createPlatformCredentialRules();
+    const credentialRuleInteractiveGuidance =
+      await this.createCredentialRuleInteractiveGuidance();
+    credentialRules.push(credentialRuleInteractiveGuidance);
+
+    return this.authorizationPolicyService.appendCredentialAuthorizationRules(
+      authorization,
+      credentialRules
+    );
   }
 
   private async propagateAuthorizationToChildEntities(
@@ -184,11 +218,7 @@ export class PlatformAuthorizationService {
     return storageAuthorization;
   }
 
-  private async appendCredentialRulesInteractiveGuidance(
-    authorization: IAuthorizationPolicy
-  ): Promise<IAuthorizationPolicy> {
-    const newRules: IAuthorizationPolicyRuleCredential[] = [];
-
+  private async createCredentialRuleInteractiveGuidance(): Promise<IAuthorizationPolicyRuleCredential> {
     const criterias: ICredentialDefinition[] = [];
     // Assign all users that are beta tester
     const betaTesterUser: ICredentialDefinition = {
@@ -213,15 +243,94 @@ export class PlatformAuthorizationService {
       this.authorizationPolicyService.createCredentialRule(
         [AuthorizationPrivilege.ACCESS_INTERACTIVE_GUIDANCE],
         criterias,
-        `${CREDENTIAL_RULE_TYPES_PLATFORM_ACCESS_GUIDANCE}2`
+        CREDENTIAL_RULE_TYPES_PLATFORM_ACCESS_GUIDANCE
       );
     interactiveGuidanceRule.cascade = false;
+    return interactiveGuidanceRule;
+  }
 
-    newRules.push(interactiveGuidanceRule);
+  private createPlatformCredentialRules(): IAuthorizationPolicyRuleCredential[] {
+    const credentialRules: IAuthorizationPolicyRuleCredential[] = [];
 
-    return this.authorizationPolicyService.appendCredentialAuthorizationRules(
-      authorization,
-      newRules
+    // Allow global admins to manage global privileges, access Platform mgmt
+    const globalAdminNotInherited =
+      this.authorizationPolicyService.createCredentialRuleUsingTypesOnly(
+        [AuthorizationPrivilege.GRANT_GLOBAL_ADMINS],
+        [AuthorizationCredential.GLOBAL_ADMIN],
+        CREDENTIAL_RULE_TYPES_PLATFORM_GRANT_GLOBAL_ADMINS
+      );
+    globalAdminNotInherited.cascade = false;
+    credentialRules.push(globalAdminNotInherited);
+
+    // Allow global admin Spaces to access Platform mgmt
+    const platformAdmin =
+      this.authorizationPolicyService.createCredentialRuleUsingTypesOnly(
+        [AuthorizationPrivilege.PLATFORM_ADMIN],
+        [
+          AuthorizationCredential.GLOBAL_ADMIN,
+          AuthorizationCredential.GLOBAL_ADMIN_SPACES,
+          AuthorizationCredential.GLOBAL_ADMIN_COMMUNITY,
+        ],
+        CREDENTIAL_RULE_TYPES_PLATFORM_ADMINS
+      );
+    platformAdmin.cascade = false;
+    credentialRules.push(platformAdmin);
+
+    // Allow all registered users to query non-protected user information
+    const userNotInherited =
+      this.authorizationPolicyService.createCredentialRuleUsingTypesOnly(
+        [AuthorizationPrivilege.READ_USERS],
+        [AuthorizationCredential.GLOBAL_REGISTERED],
+        CREDENTIAL_RULE_TYPES_PLATFORM_READ_REGISTERED
+      );
+    userNotInherited.cascade = false;
+    credentialRules.push(userNotInherited);
+
+    const createOrg =
+      this.authorizationPolicyService.createCredentialRuleUsingTypesOnly(
+        [AuthorizationPrivilege.CREATE_ORGANIZATION],
+        [
+          AuthorizationCredential.SPACE_ADMIN,
+          AuthorizationCredential.CHALLENGE_ADMIN,
+        ],
+        CREDENTIAL_RULE_TYPES_PLATFORM_ANY_ADMIN
+      );
+    createOrg.cascade = false;
+    credentialRules.push(createOrg);
+
+    const admin =
+      this.authorizationPolicyService.createCredentialRuleUsingTypesOnly(
+        [AuthorizationPrivilege.ADMIN],
+        [
+          AuthorizationCredential.GLOBAL_ADMIN,
+          AuthorizationCredential.GLOBAL_ADMIN_SPACES,
+          AuthorizationCredential.GLOBAL_ADMIN_COMMUNITY,
+          AuthorizationCredential.SPACE_ADMIN,
+          AuthorizationCredential.CHALLENGE_ADMIN,
+          AuthorizationCredential.OPPORTUNITY_ADMIN,
+          AuthorizationCredential.ORGANIZATION_ADMIN,
+        ],
+        CREDENTIAL_RULE_TYPES_PLATFORM_ANY_ADMIN
+      );
+    admin.cascade = false;
+    credentialRules.push(admin);
+
+    return credentialRules;
+  }
+
+  private createPlatformPrivilegeRules(): AuthorizationPolicyRulePrivilege[] {
+    const privilegeRules: AuthorizationPolicyRulePrivilege[] = [];
+
+    const createPrivilege = new AuthorizationPolicyRulePrivilege(
+      [
+        AuthorizationPrivilege.CREATE_SPACE,
+        AuthorizationPrivilege.CREATE_ORGANIZATION,
+      ],
+      AuthorizationPrivilege.CREATE,
+      POLICY_RULE_PLATFORM_CREATE
     );
+    privilegeRules.push(createPrivilege);
+
+    return privilegeRules;
   }
 }
